Refine diffusion acceleration policy config
