+++ python/sglang/multimodal_gen/runtime/optimization/__init__.py
@@ -1,1 +1,2 @@
 # SPDX-License-Identifier: Apache-2.0
+"""Diffusion runtime optimization policies and autotune controllers."""
